cdk/dev_function/requirements.txt
jsonpickle==1.3
aws-xray-sdk==2.4.3
geoip2>=3.0.0
python-dotenv>=0.19.2